+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPageDefault/EnrollmentPageDefault.types.js
@@ -26,6 +26,8 @@ export type Props = {|
     widgetEffects: ?WidgetEffects,
     hideWidgets: HideWidgets,
     onDelete: () => void,
+    onViewAll: (stageId: string) => void,
+    onCreateNew: (stageId: string) => void,
     onEventClick: (eventId: string, stageId: string) => void,
 |};