+++ src/core_modules/capture-core/metaDataStoreLoaders/programs/quickStoreOperations/types/apiPrograms.types.js
@@ -109,6 +109,7 @@ export type apiProgramStage = {
     dataEntryForm?: ?apiDataEntryForm,
     programStageSections?: ?Array<apiProgramStageSections>,
     programStageDataElements?: ?Array<apiProgramStageDataElement>,
+    hideDueDate?: boolean
 };
 
 type apiProgramTrackedEntityAttribute = {