+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+# [1.44.0](https://github.com/dhis2/capture-app/compare/v1.43.2...v1.44.0) (2021-08-20)
+
+
+### Features
+
+* [DHIS2-11615] add plus icon on new event button ([#1991](https://github.com/dhis2/capture-app/issues/1991)) ([64507c1](https://github.com/dhis2/capture-app/commit/64507c1cb7b7e654322cf455940f4598a5cbd90a))
+
+## [1.43.2](https://github.com/dhis2/capture-app/compare/v1.43.1...v1.43.2) (2021-08-20)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([6964c2b](https://github.com/dhis2/capture-app/commit/6964c2b9d3e73d044551e8b8b5a8f7cc67a843a0))
+
+## [1.43.1](https://github.com/dhis2/capture-app/compare/v1.43.0...v1.43.1) (2021-08-19)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([9c2b7db](https://github.com/dhis2/capture-app/commit/9c2b7dbdb0088c7663e9a9f7c89f68b22f82e55f))
+
+# [1.43.0](https://github.com/dhis2/capture-app/compare/v1.42.0...v1.43.0) (2021-08-18)
+
+
+### Features
+
+* [DHIS2-11593] display due date column based on condition ([#1972](https://github.com/dhis2/capture-app/issues/1972)) ([8ff563c](https://github.com/dhis2/capture-app/commit/8ff563cec8f1c396af843e2cbbe8b7a3895a573e))
+
+# [1.42.0](https://github.com/dhis2/capture-app/compare/v1.41.1...v1.42.0) (2021-08-18)
+
+
+### Features
+
+* [DHIS2-11572] implement navigation from overview page ([#1944](https://github.com/dhis2/capture-app/issues/1944)) ([84ee081](https://github.com/dhis2/capture-app/commit/84ee0812b42e8f31599e5ec3364cba28fd364017)), closes [#1870](https://github.com/dhis2/capture-app/issues/1870) [#1875](https://github.com/dhis2/capture-app/issues/1875)
+* [DHIS2-11594] implement comments in stages&events page ([#1971](https://github.com/dhis2/capture-app/issues/1971)) ([13996b9](https://github.com/dhis2/capture-app/commit/13996b9d89092d75aa19cf2ac8c40f15b3f334da))
+
 ## [1.41.1](https://github.com/dhis2/capture-app/compare/v1.41.0...v1.41.1) (2021-08-17)